Added instructions for FreeBSD operating System
added the instructions for the latest Hugo package though this may require constant monitoring to check if it is still the latest version and to check it. It would be better if they made the port such that it automatically gives the latest version ie. pkg install opencart as opposed to specifying the version although only for the sake of ease of instructions.